Make XLL non-losslessness warning more quiet.
Don't return WXLLLOSSY code on the first residual encoded frame. It was
confusing to users, there is nothing that can be done to correct it, and
moreover it was not entirely honest (in fact, many initial frames can be
lossy depending on how DTS core was encoded, not just one). Just drop
this warning in such situation.

Also correct an unusual condition when lossy core output would be forced
even though there are no residual encoded channel sets.